Physics/SAT_Contacts (#49)
Added SAT with contacts

+++ Light Vox Engine/ECS/EntityManager.cpp
@@ -69,7 +69,7 @@ void EntityManager::Clear()
 Entity EntityManager::Get_Entity( size_t index ) const
 {
     if ( index >= 0 && index < entries.size() )
-        return Entity( index, entries[ index ].counter );
+        return Entity( static_cast<unsigned int>(index), entries[ index ].counter );
     return {};
 }
